third_party/blink/web_tests/svg/dynamic-updates/SVGRadialGradientElement-dom-gradientUnits-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'gradientUnits' attribute of the SVGRadialGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS radialGradientElement.getAttribute('gradientUnits') is null
PASS radialGradientElement.getAttribute('gradientUnits') is "userSpaceOnUse"
PASS successfullyParsed is true

TEST COMPLETE
